make sure XLALGPSToINT8NS() returns guint64 for the psd message.

+++ src/plugins/gstlal_whiten.c
@@ -363,7 +363,7 @@ static GstMessage *psd_message_new(GSTLALWhiten *element, REAL8FrequencySeries *
 	char units[50];
 	GstStructure *s = gst_structure_new(
 		"spectrum",
-		"timestamp", G_TYPE_UINT64, XLALGPSToINT8NS(&psd->epoch),
+		"timestamp", G_TYPE_UINT64, (guint64) XLALGPSToINT8NS(&psd->epoch),
 		"delta-f", G_TYPE_DOUBLE, psd->deltaF,
 		"sample-units", G_TYPE_STRING, XLALUnitAsString(units, sizeof(units), &psd->sampleUnits),
 		"magnitude", G_TYPE_VALUE_ARRAY, va,
